# backend/edge-worker/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "A Whittle Wandering API",
    "version": "3.0.0",
    "description": "Cloudflare Worker API surface for AWhittleWandering. This schema is intended for Cloudflare API Shield Schema Validation (OpenAPI v3.0.x; no external refs)."
  },
  "servers": [
    {
      "url": "https://api.awhittlewandering.com"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/status": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/auth/token": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Issue admin JWT",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "503": {
            "description": "Not configured"
          }
        }
      }
    },
    "/api/v1/admin/ingest": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Manual data ingestion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "207": {
            "description": "Multi-Status"
          },
          "503": {
            "description": "Not configured"
          }
        }
      }
    },
    "/api/v1/admin/cache/clear": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Clear caches",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/admin/data/{resource}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin data export (allowlisted)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "vehicles",
                "journeys",
                "vehicle_state",
                "vehicle_state_history",
                "drives",
                "charges",
                "states_visited",
                "daily_analytics",
                "efficiency_metrics",
                "ingestion_logs",
                "api_cache",
                "media"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/admin/vehicles": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Onboard or update a vehicle with purchase date",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "vin"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "midnight-shadow"
                  },
                  "vin": {
                    "type": "string",
                    "pattern": "^[A-HJ-NPR-Z0-9]{17}$",
                    "example": "5YJYGDEE5LF027324"
                  },
                  "display_name": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Midnight Shadow"
                  },
                  "model": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Model Y"
                  },
                  "year": {
                    "type": "integer",
                    "minimum": 2012,
                    "example": 2023
                  },
                  "color": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Midnight Silver Metallic"
                  },
                  "purchased_date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "example": "2023-03-15",
                    "description": "Date vehicle was purchased/delivered (YYYY-MM-DD). Used as start boundary for historical data imports."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated"
          },
          "201": {
            "description": "Vehicle created"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}